feat: add rpc to update onboarding on both stripe_users and service_providers
supabase does not support transactions out of the box, so we are using a
Postgres Function as a workaround since it implicitly runs both updates
as a single transaction.

We'll use this behaviour to ensure that either both tables are updated,
or both are not updated, and we don't end up with an inconsistent state.