migrations: add index on task_category_id
This commit renames the task_categories_id field on the tasks table
to task_category_id and adds an index on it. This directly updates the
original migration file as I plan to squash all commits in this branch.

I assumed that an index would automatically get added on the foreign key,
but it looks like this is wrong. We'll have to update the FKs on the
other tables as well.